airflow/requirements.txt
clickhouse-driver
clickhouse-connect
psycopg2-binary
requests
kafka-python